Lookup gene sequences from NCBI

Look for sequences of a particular gene for a list of species/isolates/genes from the Genbank nucleotide database.

get_isolate_seqs(species, genes, isolates = NULL,
  extract_features = FALSE, gene_name_in_feature = TRUE,
  flanking = c(0, 0), db = "nucleotide", pause = 0.5, ...)

Arguments

species
The names of species to look up.

genes
The names of the genes to look up.

isolates
The names of isolates to look up. Must be the same length as species if used.

extract_features
If TRUE, return the sequence for each feature in the sequence annotation instead of the whole sequence.

gene_name_in_feature
If TRUE, only return features that have one of the gene names somewhere in their description. Only has an effect if extract_features is TRUE.

flanking
A vector of length 2. The number of base pairs before and after the target gene to include in the sequence returned. If the flanking sequence is not available, the sequence will be considered incomplete.

db
The name of the NCBI database to query. Only tested with "nucleotide", but a few others might work.

pause
The number of seconds to pause between each query. This avoids annoying NCBI and having them block you IP address. Should be at least 0.35 seconds if you dont have an NCBI API key and at least 0.1 seconds if you do.

...
Additional terms to add to the search request for each species/isolate, see NCBI documentation for a complete list: http://www.ncbi.nlm.nih.gov/books/NBK25499/#chapter4_ESearch

Value

A tibble (a type of data.frame) with the following columns, depending on settings:

species: The species search term associated with the result

isolate: The isolate search term associated with the result

query: The query used to search for sequences on NCBI

acc: The Genbank accession number

start: The index of the first base pair of the target gene in the original sequence

end: The index of the last base pair of the target gene in the original sequence

feature: The type of locus

type: The type of annotation for the sequence

name: The name of the locus

complete: If the locus was complete in the original sequence, according to the annotation

seq: The whole sequence the gene was found in

header: The name of the overall sequence the gene was found in

length: The length of the whole sequence

flank_start: The index of the first base pair of the target gene plus the flanking region in the original sequence

flank_end: The index of the last base pair of the target gene plus the flanking region in the original sequence

flank_complete: If the locus plus flanking region was complete in the original sequence

gene_seq: The target gene sequence plus the flanking region

gene_length: The length of the target gene plus the flanking region

Examples

if (FALSE) {

# Search for the whole seqeunces for with P. infestans Cox I
get_isolate_seqs(species = c("Phytophthora infestans"),
                 genes = c("cox I", "cox 1", "cox1", "coxI", "cytochrome oxidase I", "cytochrome oxidase 1"),
                 retmax = 100)

# Search for the just the gene sequence for  P. infestans Cox I
get_isolate_seqs(species = c("Phytophthora infestans"),
                 genes = c("cox I", "cox 1", "cox1", "coxI", "cytochrome oxidase I", "cytochrome oxidase 1"),
                 retmax = 100,
                 extract_features = TRUE)

# Search for all the gene sequences in whole sequences that contain P. infestans Cox I
get_isolate_seqs(species = c("Phytophthora infestans"),
                 genes = c("cox I", "cox 1", "cox1", "coxI", "cytochrome oxidase I", "cytochrome oxidase 1"),
                 retmax = 100,
                 extract_features = TRUE,
                 gene_name_in_feature = FALSE)

# Search for whole sequences for P. infestans Cox I for just some isolates
get_isolate_seqs(species = c("Phytophthora infestans", "Phytophthora infestans", "Phytophthora infestans"),
                 isolates = c("44", "580", "180"),
                 genes = c("cox I", "cox 1", "cox1", "coxI", "cytochrome oxidase I", "cytochrome oxidase 1"))

# Search for just the gene sequences for P. infestans Cox I for just some isolates
get_isolate_seqs(species = c("Phytophthora infestans", "Phytophthora infestans", "Phytophthora infestans"),
                 isolates = c("44", "580", "180"),
                 genes = c("cox I", "cox 1", "cox1", "coxI", "cytochrome oxidase I", "cytochrome oxidase 1"),
                 extract_features = TRUE)

# Search for P infestans ITS with flanking regions and subset for complete results
result <- get_isolate_seqs(species = c("Phytophthora"),
                           genes = c("internal transcribed spacer"),
                           retmax = 300,
                           extract_features = TRUE,
                           flanking = c(300, 300))
result[result$complete, ]
result[result$flank_complete, ]

}